fix: Consolidate imports in dashboard components for improved organization
