+++ apps/frontend/app/components/dashboard/forms/review-entity-form.tsx
@@ -33,7 +33,7 @@ import { useState } from "react";
 import { Form } from "react-router";
 import { match } from "ts-pattern";
 import { withQuery } from "ufo";
-import { convertThreePointSmileyToDecimal } from "~/components/dashboard/utils";
+import { convertThreePointSmileyToDecimal } from "~/components/dashboard";
 import {
 	ThreePointSmileyRating,
 	convertDecimalToThreePointSmiley,
