+++ apps/frontend/app/components/dashboard/index.tsx
@@ -1,0 +1,11 @@
+export * from "./error-boundary";
+export * from "./forms/add-entity-to-collections-form";
+export * from "./forms/create-measurement-form";
+export * from "./forms/review-entity-form";
+export * from "./modals/layout-modals";
+export * from "./modals/metadata-progress-update-forms";
+export * from "./navigation/footer";
+export * from "./navigation/links-group";
+export * from "./navigation/navigation-config";
+export * from "./types";
+export * from "./utils";
